Refactor: admin용 장소 등록 페이지 제거 및 기존 장소 등록에 자동 카테고리 기능 추가
