+++ types/corePlace.ts
@@ -18,12 +18,147 @@ export const PLACE_CATEGORIES = [
   { value: 'medical', label: '의료' },
   { value: 'gas_station', label: '주유소' },
   { value: 'parking', label: '주차장' },
-
   { value: 'etc', label: '기타' },
 ] as const;
 
 export type PlaceCategory = (typeof PLACE_CATEGORIES)[number]['value'];
 // → 'restaurant' | 'cafe' | 'hotel' | ... | 'etc'
+
+export const TYPE_MAP: Record<string, PlaceCategory> = {
+  airport: 'airport',
+  train_station: 'traffic',
+  subway_station: 'traffic',
+  bus_station: 'traffic',
+  transit_station: 'traffic',
+  bus_stop: 'traffic',
+  ferry_terminal: 'terminal',
+  restaurant: 'restaurant',
+  meal_takeaway: 'restaurant',
+  meal_delivery: 'restaurant',
+  fast_food_restaurant: 'restaurant',
+  japanese_restaurant: 'restaurant',
+  chinese_restaurant: 'restaurant',
+  korean_restaurant: 'restaurant',
+  american_restaurant: 'restaurant',
+  italian_restaurant: 'restaurant',
+  french_restaurant: 'restaurant',
+  seafood_restaurant: 'restaurant',
+  brunch_restaurant: 'restaurant',
+  dutch_restaurant: 'restaurant',
+  hamburgur_restaurant: 'restaurant',
+  halal_restaurant: 'restaurant',
+  mexican_restaurant: 'restaurant',
+  asian_restaurant: 'restaurant',
+  dim_sum_restaurant: 'restaurant',
+  steak_house: 'restaurant',
+  buffet_restaurant: 'restaurant',
+  fine_dining_restaurant: 'restaurant',
+  brewery: 'restaurant',
+  deli: 'restaurant',
+  wine_bar: 'restaurant',
+  winery: 'restaurant',
+  cafe: 'cafe',
+  cafteria: 'cafe',
+  dessert_shop: 'cafe',
+  cake_shop: 'cafe',
+  cat_cafe: 'cafe',
+  dog_cafe: 'cafe',
+  ice_cream_shop: 'cafe',
+  dessert_restaurant: 'cafe',
+  coffee_shop: 'cafe',
+  tea_house: 'cafe',
+  bakery: 'cafe',
+  bagle_shop: 'cafe',
+  lodging: 'hotel',
+  hotel: 'hotel',
+  motel: 'hotel',
+  bed_and_breakfast: 'hotel',
+  resort_hotel: 'hotel',
+  shopping_mall: 'shopping',
+  department_store: 'shopping',
+  clothing_store: 'shopping',
+  shoe_store: 'shopping',
+  jewelry_store: 'shopping',
+  book_store: 'shopping',
+  electronics_store: 'shopping',
+  furniture_store: 'shopping',
+  home_goods_store: 'shopping',
+  market: 'shopping',
+  outlet_mall: 'shopping',
+  food_store: 'shopping',
+  gift_shop: 'shopping',
+  tea_store: 'shopping',
+  cosmetics_store: 'shopping',
+  convenience_store: 'convenience',
+  supermarket: 'shopping',
+  grocery_store: 'shopping',
+  atm: 'atm',
+  bank: 'atm',
+  museum: 'culture',
+  art_museum: 'culture',
+  art_gallery: 'culture',
+  art_studio: 'culture',
+  church: 'culture',
+  place_of_worship: 'culture',
+  historical_landmark: 'landmark',
+  monument: 'culture',
+  cultural_landmark: 'landmark',
+  performing_arts_theater: 'culture',
+  opera_house: 'culture',
+  concert_hall: 'culture',
+  history_museum: 'culture',
+  mosque: 'culture',
+  amusement_park: 'leisure',
+  water_park: 'leisure',
+  marina: 'leisure',
+  zoo: 'leisure',
+  aquarium: 'leisure',
+  night_club: 'leisure',
+  casino: 'leisure',
+  bowling_alley: 'leisure',
+  movie_theater: 'leisure',
+  spa: 'leisure',
+  park: 'nature',
+  beach: 'nature',
+  campground: 'nature',
+  national_park: 'nature',
+  hiking_area: 'nature',
+  lake: 'nature',
+  mountain: 'nature',
+  garden: 'nature',
+  waterfall: 'nature',
+  river: 'nature',
+  woods: 'nature',
+  stadium: 'sports',
+  gym: 'sports',
+  sports_complex: 'sports',
+  golf_course: 'sports',
+  ski_resort: 'sports',
+  swimming_pool: 'sports',
+  hospital: 'medical',
+  pharmacy: 'medical',
+  doctor: 'medical',
+  dentist: 'medical',
+  clinic: 'medical',
+  massage: 'medical',
+  sauna: 'medical',
+  school: 'education',
+  university: 'education',
+  library: 'education',
+  gas_station: 'gas_station',
+  rest_stop: 'traffic',
+  electric_vehicle_charging_: 'traffic',
+  parking: 'parking',
+  parking_lot: 'parking',
+  parking_garage: 'parking',
+  tourist_attraction: 'landmark',
+  point_of_interest: 'landmark',
+  landmark: 'landmark',
+};
+
+export function mapPrimaryTypeToCategory(primaryType: string): PlaceCategory {
+  return TYPE_MAP[primaryType] ?? 'etc';
+}
 
 export interface CreatePlacePayload {
   google_place_id: string;
